begin moving print statements from `ptr` to `io`
